Reshape crashfix and added caching at other places
